dot/types, lib/babe/types: migrate babe header to dot types (#771)
* dot/types, lib/babe: migrate babe types to dot types
