+++ dot/core/runtime_test.go
@@ -28,7 +28,6 @@ import (
 	"github.com/ChainSafe/gossamer/lib/runtime"
 	"github.com/ChainSafe/gossamer/lib/transaction"
 	"github.com/ChainSafe/gossamer/lib/trie"
-	"github.com/ChainSafe/gossamer/tests"
 
 	"github.com/stretchr/testify/require"
 )
@@ -41,7 +40,7 @@ func TestRetrieveAuthorityData(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	err = tt.Put(tests.AuthorityDataKey, value)
+	err = tt.Put(TestAuthorityDataKey, value)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -102,7 +101,7 @@ func TestCheckForRuntimeChanges(t *testing.T) {
 	require.Nil(t, err)
 
 	pubkey := kp.Public().Encode()
-	err = tt.Put(tests.AuthorityDataKey, append([]byte{4}, pubkey...))
+	err = tt.Put(TestAuthorityDataKey, append([]byte{4}, pubkey...))
 	require.Nil(t, err)
 
 	ks := keystore.NewKeystore()
